Added more messages, better PATHEXT save
